Initial version of victoria_trex_ros_test
This version will receive messages from a ROS subscriber to control the motors via the TRex motor controller, and will send messages using a ROS publisher with information from the motor encoders. It is not meant to be compatible with any existing ROS navigation stack, but demonstrates the connectivity to ROS using ros_serial. It also exercises functionality expected to be setup using Victoria's Teensy circuit, separate from the main computer.